refactoring loaders.py to have consistent handling of xobs and (now) thetafid.

+++ ili/dataloaders/loaders.py
@@ -48,16 +48,30 @@ class NumpyLoader(_BaseLoader):
     """A class for loading in-memory data using numpy arrays.
 
     Args:
-        x (np.array): Array of data of shape (Ndata, \*data.shape)
-        theta (np.array): Array of parameters of shape (Ndata, \*theta.shape)
+        x (np.array): Array of training data of
+            shape (Ndata, \*data.shape)
+        theta (np.array): Array of training parameters of
+            shape (Ndata, \*parameters.shape)
+        xobs (Optional[np.array]): Array of observed data of
+            shape (\*data.shape). Defaults to None.
+        thetafid (Optional[np.array]): Array of fiducial
+            parameters of shape (\*parameters.shape). Defaults to None.
     """
 
-    def __init__(self, x, theta) -> None:
+    def __init__(
+        self,
+        x: np.array,
+        theta: np.array,
+        xobs: Optional[np.array] = None,
+        thetafid: Optional[np.array] = None
+    ) -> None:
         self.x = x
         self.theta = theta
         if len(self.x) != len(self.theta):
             raise Exception(
                 "Stored data and parameters are not of same length.")
+        self.xobs = xobs
+        self.thetafid = thetafid
 
     def __len__(self) -> int:
         """Returns the total number of data points in the dataset
@@ -68,7 +82,7 @@ class NumpyLoader(_BaseLoader):
         return len(self.x)
 
     def get_all_data(self) -> np.array:
-        """Returns all the loaded data
+        """Returns all the loaded data for training
 
         Returns:
             np.array: data
@@ -76,12 +90,29 @@ class NumpyLoader(_BaseLoader):
         return self.x
 
     def get_all_parameters(self):
-        """Returns all the loaded parameters
+        """Returns all the loaded parameters for training
 
         Returns:
             np.array: parameters
         """
         return self.theta
+
+    def get_obs_data(self) -> np.array:
+        """Returns the observed data
+
+        Returns:
+            np.array: data
+        """
+        return self.xobs
+
+    def get_fid_parameters(self):
+        """Returns the fiducial parameters which we expect the
+        observed data to resemble
+
+        Returns:
+            np.array: parameters
+        """
+        return self.thetafid
 
 
 class StaticNumpyLoader(NumpyLoader):
@@ -89,19 +120,41 @@ class StaticNumpyLoader(NumpyLoader):
 
     Args:
         in_dir (str): path to the location of stored data
-        x_file (str): filename of the stored data
-        theta_file (str): filename of the stored parameters
+        x_file (str): filename of the stored training data
+        theta_file (str): filename of the stored training parameters
+        xobs_file (Optional[str]): filename used for observed x values
+        thetafid_file (Optional[str]): filename used for fiducial parameters
     """
 
-    def __init__(self, in_dir: str, x_file: str, theta_file: str):
+    def __init__(
+        self,
+        in_dir: str,
+        x_file: str,
+        theta_file: str,
+        xobs_file: Optional[str] = None,
+        thetafid_file: Optional[str] = None
+    ) -> None:
         self.in_dir = Path(in_dir)
         self.x_path = self.in_dir / x_file
         self.theta_path = self.in_dir / theta_file
 
+        # Load stored data (if specified)
         x = np.load(self.x_path)
         theta = np.load(self.theta_path)
+        if xobs_file is None:
+            self.xobs_path = None
+            xobs = None
+        else:
+            self.xobs_path = self.in_dir / xobs_file
+            xobs = np.load(self.xobs_path)
+        if thetafid_file is None:
+            self.thetafid_path = None
+            thetafid = None
+        else:
+            self.thetafid_path = self.in_dir / thetafid_file
+            thetafid = np.load(self.thetafid_path)
 
-        super().__init__(x=x, theta=theta)
+        super().__init__(x=x, theta=theta, xobs=xobs, thetafid=thetafid)
 
 
 class SummarizerDatasetLoader(_BaseLoader):
@@ -211,7 +264,7 @@ class SummarizerDatasetLoader(_BaseLoader):
         return theta[param_names].values
 
 
-class SBISimulator(_BaseLoader):
+class SBISimulator(NumpyLoader):
     """
     Class to run simulations of data and parameters and save
     results to numpy files. Only works for sbi backend.
@@ -219,40 +272,68 @@ class SBISimulator(_BaseLoader):
     Args:
         in_dir (str): path to the location of stored data
         xobs_file (str): filename used for observed x values
-        thetaobs_file (str): filename used for observed parameters
-        out_dir (str): path to the location where to save  data
-        x_file (str): filename to use to store data
-        theta_file (str): filename to use to store parameters
         num_simulations (int): number of simulations to run at each call
         simulator (callable): function taking the parameters as an
             argument and returns data
+        save_simulated (Optional[bool]): whether to save simulated data.
+            Concatenates to previous data if True. Defaults to False.
+        x_file (Optional[str]): filename of the stored first-round
+            training data
+        theta_file (Optional[str]): filename of the stored first-round
+            training parameters
+        thetafid_file (Optional[str]): filename used for fiducial parameters
     """
 
     def __init__(
-            self,
-            in_dir: str,
-            xobs_file: str,
-            thetaobs_file: str,
-            out_dir: str,
-            x_file: str,
-            theta_file: str,
-            num_simulations: int,
-            simulator: Optional[callable] = None,
+        self,
+        in_dir: str,
+        xobs_file: str,
+        num_simulations: int,
+        simulator: Optional[callable] = None,
+        save_simulated: Optional[bool] = False,
+        x_file: Optional[str] = None,
+        theta_file: Optional[str] = None,
+        thetafid_file: Optional[str] = None,
     ):
         self.in_dir = Path(in_dir)
         self.xobs_path = self.in_dir / xobs_file
-        self.thetaobs_path = self.in_dir / thetaobs_file
-        self.out_dir = Path(out_dir)
-        self.x_path = self.out_dir / x_file
-        self.theta_path = self.out_dir / theta_file
         self.num_simulations = num_simulations
         self.simulator = simulator
+        self.save_simulated = save_simulated
 
-        self.xobs = np.load(self.xobs_path)
-        self.thetaobs = np.load(self.thetaobs_path)
+        # If save_simulated, check that x_file and theta_file are specified
+        if save_simulated and (x_file is None or theta_file is None):
+            raise Exception(
+                "If save_simulated is True, x_file and theta_file must be "
+                "specified."
+            )
 
-        self.theta = None
-        self.x = None
+        # Load stored data (if specified)
+        xobs = np.load(self.xobs_path)
+        x = None
+        theta = None
+        thetafid = None
+        if x_file is None:
+            self.x_path = None
+        else:
+            self.x_path = self.in_dir / x_file
+            if self.x_path.exists():
+                x = np.load(self.x_path)
+        if theta_file is None:
+            self.theta_path = None
+        else:
+            self.theta_path = self.in_dir / theta_file
+            if self.theta_path.exists():
+                theta = np.load(self.theta_path)
+        if thetafid_file is None:
+            self.thetafid_path = None
+        else:
+            self.thetafid_path = self.in_dir / thetafid_file
+            thetafid = np.load(self.thetafid_path)
+
+        super().__init__(
+            x=x, theta=theta, xobs=xobs, thetafid=thetafid
+        )
 
     def __len__(self) -> int:
         """Returns the total number of data points produced when called
@@ -285,46 +366,18 @@ class SBISimulator(_BaseLoader):
         theta = proposal.sample((self.num_simulations,)).cpu()
         x = simulate_in_batches(self.simulator, theta)
         theta, x = theta.numpy(), x.numpy()
+
+        # Save simulated data (concatenates to previous data)
         if self.theta is None or self.x is None:
             self.theta, self.x = theta, x
         else:
             self.theta = np.concatenate((self.theta, theta))
             self.x = np.concatenate((self.x, x))
-        np.save(self.theta_path, self.theta)
-        np.save(self.x_path, self.x)
+        if self.save_simulated:
+            np.save(self.theta_path, self.theta)
+            np.save(self.x_path, self.x)
+
         return theta, x
-
-    def get_obs_data(self) -> np.array:
-        """Returns the observed data
-
-        Returns:
-            np.array: data
-        """
-        return self.xobs
-
-    def get_obs_parameters(self):
-        """Returns the observed parameters
-
-        Returns:
-            np.array: parameters
-        """
-        return self.thetaobs
-
-    def get_all_data(self) -> np.array:
-        """Returns all the loaded data
-
-        Returns:
-            np.array: data
-        """
-        return self.x
-
-    def get_all_parameters(self):
-        """Returns all the loaded parameters
-
-        Returns:
-            np.array: parameters
-        """
-        return self.theta
 
 
 # TODO: Add loaders which load dynamically from many files, so
